reorganising E2E tests where SDK validation and functionality tests are in the same file
